requirements.txt
python-dotenv
pyyaml
streamlit
crewai
crewai-tools
langchain-community
langchain-openai
langchain-text-splitters 
langchain-chroma 
langchain-openai
chromadb
qdrant-client
markitdown
pypdf
openai
tiktoken
numpy
tqdm
